backend/requirements.txt
# backend/requirements.txt

# Core Flask Framework
Flask>=3.0.0
Flask-SQLAlchemy>=3.1.1
Flask-JWT-Extended>=4.6.0
Flask-CORS>=4.0.0

# Database
SQLAlchemy>=2.0.23
# sqlite3  # Built into Python

# Azure OpenAI and AI Services
openai>=1.51.2
azure-identity>=1.15.0

# HTTP Requests and Web
requests>=2.31.0
urllib3>=2.1.0

# Data Processing and Validation
pydantic>=2.5.2
marshmallow>=3.20.2

# Security and Authentication
Werkzeug>=3.0.1
PyJWT>=2.8.0
bcrypt>=4.1.2

# Background Tasks and Concurrency
threading  # Built into Python
concurrent.futures  # Built into Python

# Utilities
python-dotenv>=1.0.0
click>=8.1.7
python-dateutil>=2.8.2

# Development and Testing
pytest>=7.4.3
pytest-flask>=1.3.0
pytest-cov>=4.1.0

# Logging and Monitoring
structlog>=23.2.0

# Optional: Enhanced functionality
# selenium>=4.15.2  # For web scraping (optional)
# beautifulsoup4>=4.12.2  # For HTML parsing (optional)
# webdriver-manager>=4.0.1  # For managing web drivers (optional)

# Production WSGI Server (optional)
gunicorn>=21.2.0

# Caching (optional)
# redis>=5.0.1
# Flask-Caching>=2.1.0

# Email (optional)
# Flask-Mail>=0.9.1

# Rate Limiting (optional)  
# Flask-Limiter>=3.5.0